Pass the new Front response back to state to help redux to update the Front app with recent data and not going blank.
